rename: lotto 패키지 추가
- Lotto와 Lottos를 lotto 패키지로 묶는다.
