+++ src/main/java/org/duckstudy/controller/LottoController.java
@@ -3,8 +3,8 @@ package org.duckstudy.controller;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import org.duckstudy.model.Lotto;
-import org.duckstudy.model.Lottos;
+import org.duckstudy.model.lotto.Lotto;
+import org.duckstudy.model.lotto.Lottos;
 import org.duckstudy.model.Price;
 import org.duckstudy.view.InputView;
 import org.duckstudy.view.OutputView;
